pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm>=8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-workspace"
dynamic = ["version"]
authors = [
    {name = "Marcus Jellinghaus"},
]
description = "A simple Model Context Protocol (MCP) server with file operation tools"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["mcp", "server", "filesystem", "claude", "ai", "assistant", "file-operations"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pathspec>=0.12.1",
    "igittigitt>=2.1.5",
    "mcp>=1.3.0",
    "GitPython>=3.1.0",
    "mcp-coder-utils",
    "PyGithub>=1.59.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-xdist>=3.6.1",
    "pylint>=3.3.3",
    "black>=24.10.0",
    "isort>=5.13.2",
    "mypy>=1.13.0",
    "types-requests>=2.31.0",
    # "mcp-tools-py",  # TODO: uncomment once mcp-tools-py is published on PyPI
    # "mcp-coder",  # TODO: uncomment once mcp-coder is published on PyPI
]
architecture = [
    "import-linter>=2.0",
    "tach>=0.6.0",
    "pycycle>=0.0.8",
    "vulture>=2.14",
]
config = [
    "mcp-config",
]

[project.scripts]
mcp-workspace = "mcp_workspace.main:main"

[project.urls]
"Homepage" = "https://github.com/MarcusJellinghaus/mcp-workspace"
"Bug Tracker" = "https://github.com/MarcusJellinghaus/mcp-workspace/issues"
"Documentation" = "https://github.com/MarcusJellinghaus/mcp-workspace#readme"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.setuptools_scm]
tag_regex = '^(?P<version>\d+\.\d+\.\d+)$'
# Version is automatically determined from git tags
# Tag format: 0.1.0 -> version 0.1.0
# Between releases: 0.1.0.devN+gHASH (includes commit count and hash)

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["src"]
addopts = ["--strict-markers", "--tb=short", "-n", "auto", "-p", "no:tach"]
markers = [
    "git_integration: File system git operations (repos, commits)",
    "github_integration: GitHub API access (requires GITHUB_TOKEN)",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88
float_to_top = true

[tool.mcp-coder.install-from-github]
# Installed WITH deps (leaves — picks up new external deps)
packages = [
    "mcp-config-tool @ git+https://github.com/MarcusJellinghaus/mcp-config.git",
    "mcp-coder-utils @ git+https://github.com/MarcusJellinghaus/mcp-coder-utils.git",
]
# Installed WITHOUT deps (depend on siblings — avoid downgrading)
packages-no-deps = [
    "mcp-tools-py @ git+https://github.com/MarcusJellinghaus/mcp-tools-py.git",
    "mcp-coder @ git+https://github.com/MarcusJellinghaus/mcp_coder.git",
]

[tool.pylint.messages_control]
# W, C, R categories disabled - only errors (E) and fatals (F) are checked
# Previously disabled individually (now covered by W category above):
# W1203: Disable lazy % formatting - project prefers f-strings for readability
# W0621: Disable redefined-outer-name - pytest fixtures standard pattern
# R0401: Enable cyclic-import detection (not enabled by default as it's in refactor category)
disable = ["W", "C", "R"]
enable = ["R0401"]

[tool.mypy]
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true
strict = true
warn_unreachable = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_decorators = true

[[tool.mypy.overrides]]
module = ["pytest.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
# Disable unreachable warnings for test files due to mypy false positives with assertions
# Disable union-attr for test files where mocked attributes are intentionally set
disable_error_code = ["unreachable", "union-attr"]
ignore_missing_imports = true
